Added conditional processing to handle excess white space characters
Added conditional processing to handle excess white space characters
between the '<' and digits in the lines with pollution data. The
characters seem to have been added during conversion from PDF to TXT
file format and were throwing off the results of the script because the
re.split bit (line 56) was, in those cases, causing the '<' (second
column) to be read as the annual emissions and the annual emissions
(third column) as the maximum hourly emissions.